sophia: support data compression, close #67
From now on Sophia is capable to use compression
for a data storage.

Unit of compression is one page (sorted region).
Keys, values and metadata are included in
compression. Compression happen in background for
the branch/compaction operations. Pages are
decompressed on a read.

It is possible to have separate compression
policy for each database.

First tests of 'test/mt_backend.test.c' gave up
to x10 win in space usage.

More compression drivers and efficient split of
compressed node files is yet to come.